+++ Comparator/Resource.h
@@ -25,13 +25,14 @@
 #define ID_MOUSEMENU_START              0x7600
 #define ID_MOUSEMENU_END                0x76ff
 #define ID_MAGNIFY                      32771
+#define ID_COMPARATOR_HELP              32772
 
 // Next default values for new objects
 // 
 #ifdef APSTUDIO_INVOKED
 #ifndef APSTUDIO_READONLY_SYMBOLS
 #define _APS_NEXT_RESOURCE_VALUE        311
-#define _APS_NEXT_COMMAND_VALUE         32772
+#define _APS_NEXT_COMMAND_VALUE         32773
 #define _APS_NEXT_CONTROL_VALUE         1004
 #define _APS_NEXT_SYMED_VALUE           310
 #endif